Add MinPixelBrightness
MinPixelBrightness defines a minimum level of gray that can be displayed
during rendering. If a color defines a component less than this value,
it is still allowed. If all components (R, G, and B) are less than this
value, the the color is changed to the level of gray defined by this
component value. This setting affects both rendering in the app
and exporting image(s).